Add Filter to GetDevice RPC

+++ juno/devices/yeelight/YeelightBulb.hh
@@ -43,7 +43,7 @@ private:
     kstd::UniquePtr<StreamSocket> m_socket;
 
 public:
-    const i32 yeelightId;
+    const u64 yeelightId;
 };
 
 }  // namespace juno
